% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakers_wide.R
\docType{data}
\name{bakers_wide}
\alias{bakers_wide}
\title{Great British Bakeoff spice scores in wide format}
\format{
A tibble with 3690 rows and 8 variables:
\describe{
\item{baker}{chr ID for beach site}
\item{cinnamon_1}{dbl Score on cinnamon test}
\item{cardomom_2}{dbl Score on cardamom test}
\item{nutmeg_3}{dbl Score on nutmeg test}

}
}
\source{
\url{https://bakeoff.netlify.app/index.html}
}
\usage{
bakers_wide
}
\description{
Great British Bakeoff spice scores in wide format
}
\keyword{datasets}
