% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beachbugs_long.R
\docType{data}
\name{beachbugs_long}
\alias{beachbugs_long}
\title{Beachbugs long}
\format{
A tibble with 66 rows and 3 variables:
\describe{
\item{year}{dbl Year}
\item{site}{chr Beach site}
\item{buglevels}{dbl Mean bacteria levels}

}
}
\source{
\url{https://www.environment.nsw.gov.au/topics/water/beaches/beachwatch-water-quality-program}
}
\usage{
beachbugs_long
}
\description{
Long format mean Enterococci bacteria levels by beach site and year between 2013 and 2018
}
\keyword{datasets}
