% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beachbugs_wide.R
\docType{data}
\name{beachbugs_wide}
\alias{beachbugs_wide}
\title{Beachbugs wide}
\format{
A tibble with 6 rows and 12 variables:
\describe{
\item{year}{dbl Year}
\item{\verb{Bondi Beach}}{dbl Mean bug levels at bondi}
\item{\verb{Bronte Beach}}{dbl Mean bug levels at bronte}
\item{\verb{Clovelly Beach}}{dbl Mean bug levels at clovelly}
\item{\verb{Coogee Beach}}{dbl Mean bug levels at coogee}
\item{\verb{Gordons Bay (East)}}{dbl Mean bug levels at gbe}
\item{\verb{Little Bay Beach}}{dbl Mean bug levels at lb}
\item{\verb{Malabar Beach}}{dbl Mean bug levels at malabar}
\item{\verb{Maroubra Beach}}{dbl Mean bug levels at maroubra}
\item{\verb{South Maroubra Beach}}{dbl Mean bug levels at smb}
\item{\verb{South Maroubra Rockpool}}{dbl Mean bug levels at smr}
\item{\verb{Tamarama Beach}}{dbl Mean bug levels at tama}
}
}
\source{
\url{https://www.environment.nsw.gov.au/topics/water/beaches/beachwatch-water-quality-program}
}
\usage{
beachbugs_wide
}
\description{
Wide format mean Enterococci bacteria levels by beach site and year between 2013 and 2018
}
\keyword{datasets}
