% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames_long.R
\docType{data}
\name{frames_long}
\alias{frames_long}
\title{Frames data in long format}
\format{
A tibble with 4725 rows and 7 variables:
#' \describe{
\item{id}{chr Participant ID}
\item{gender}{chr Participant gender}
\item{age}{dbl Participant age in years}
\item{condition}{chr condition category or property}
\item{testitem}{chr Test Items 1-7}
\item{response}{dbl Response}

}
}
\source{
\url{https://psyarxiv.com/2m83v/}
}
\usage{
frames_long
}
\description{
Frames data in long format
}
\keyword{datasets}
