% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryouwithme-package.R
\docType{package}
\name{ryouwithme-package}
\alias{ryouwithme}
\alias{ryouwithme-package}
\title{ryouwithme: RYouWithMe Data}
\description{
This package dontains dataset that will be useful to those learning R using the RYouWithMe modules from RLadies-Sydney.
}
\author{
\strong{Maintainer}: Jen Richmond \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
