% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sydneybeaches.R
\docType{data}
\name{sydneybeaches}
\alias{sydneybeaches}
\title{Enterococci bacteria levels at Sydney beaches between 2013 and 2018}
\format{
A tibble with 3690 rows and 8 variables:
\describe{
\item{BeachID}{dbl ID for beach site}
\item{Region}{chr Region}
\item{Council}{dbl City Council responsible for beach site}
\item{Site}{chr Name of the beach site}
\item{Longitude}{dbl Longitude for beach site}
\item{Latitude}{dbl Latitude for beach site}
\item{Date}{chr Date that sample was taken}
\item{\verb{Enterococci (cfu/100ml)}}{dbl Bug levels for beach site}
}
}
\source{
\url{https://www.environment.nsw.gov.au/topics/water/beaches/beachwatch-water-quality-program}
}
\usage{
sydneybeaches
}
\description{
Enterococci bacteria levels at Sydney beaches between 2013 and 2018
}
\keyword{datasets}
